Refactor inputs and outputs to be Doubles

+++ src/perceptron/OrPerceptron.kt
@@ -1,7 +1,7 @@
 package perceptron
 
 class OrPerceptron: Perceptron {
-    fun calculateOutput(input1: Int, input2: Int): Int {
+    fun calculateOutput(input1: Double, input2: Double): Double {
         return super.calculateOutput(listOf(input1, input2), listOf(1.0, 1.0),0.0)
     }
 }
